DJGPP updates for 64-bit MX
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@260 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_amplifier.c
@@ -7,7 +7,7 @@
  *
  *--------------------------------------------------------------------------
  *
- * Copyright 1999, 2001-2003, 2005 Illinois Institute of Technology
+ * Copyright 1999, 2001-2003, 2005-2006 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -331,7 +331,7 @@ mx_amplifier_default_get_parameter_handler( MX_AMPLIFIER *amplifier )
 			"by the MX driver for amplifier '%s'.",
 			mx_get_field_label_string( amplifier->record,
 						amplifier->parameter_type ),
-			amplifier->parameter_type,
+			(int) amplifier->parameter_type,
 			amplifier->record->name );
 		break;
 	}
@@ -354,7 +354,7 @@ mx_amplifier_default_set_parameter_handler( MX_AMPLIFIER *amplifier )
 			"by the MX driver for amplifier '%s'.",
 			mx_get_field_label_string( amplifier->record,
 						amplifier->parameter_type ),
-			amplifier->parameter_type,
+			(int) amplifier->parameter_type,
 			amplifier->record->name );
 		break;
 	}
